Refactor code for adding task to project
Passing the project object from show-projects module through to
add-task-form was causing issues when a task was not submitted to
project. In the case of clicking "add task" and then pressing cancel,
the next time a task was to be submitted to a project, the function
ran twice: once for the original "add task" execution and again for the
second "add task" execution.